Put the mobile sync wire on a diet
Two changes that cut what a phone downloads, both measured against this
machine's live data (P0 items 1 and 2).

Chat events crossed the wire uncompacted. `commitChatEvent` built two
envelopes: the stored one went through `compactChatEventForStorage`, the live
one handed to `eventSubscribers` — which is what the sync host pushes to phones
and web — went out whole, transformed only by inline-image redaction. The same
event was therefore multi-megabyte on live push and small after reconnect
hydration, a size change users could observe.

The deeper defect was two compaction implementations with two cap tables, which
drifted the way duplicated policy does. `tool_result.structured` was added to
the event and to neither table. On a real 8 MB transcript it had grown to
4.53 MB — 56.6% of the entire file, and ten times larger than `result`, the
field that IS capped. So this collapses both into shared/chatEventCompaction,
where a new heavy field cannot be capped on one side and forgotten on the other.

`structured` is now bounded on disk and dropped from the wire outright, along
with `toolResultMeta`. Everything ADE reads from `structured` (grep totals, bash
timeout and cwd hints, subagent enrichment) is projected into typed fields on
the same event at construction time; past that point nothing consumes it.
`structured` is not a coding key in the iOS decoder at all, and `toolResultMeta`
is written once and never read on any surface — the phone was downloading and
parsing megabytes to produce something it discarded. Removing a field no client
decodes is backward-compatible by construction, so this needs no capability
gate; anything that added or reshaped a field would.

Measured on a real thread, replaying every event through the new wire path:
7.99 MB → 3.46 MB overall, and tool_result 5.10 MB → 0.57 MB across 356
records. That transcript is already storage-compacted, so the live-push saving
is larger than the 2.3x shown.

`pull_request_snapshots` no longer rides the mobile changeset pump. It was
10.65 MB of a 26.8 MB synced project database — 39.7%, 258 rows averaging 42 KB,
one `files_json` at 1.58 MB — for data the phone already fetched a second time
itself. iOS reads the table in exactly one SELECT, the per-PR detail query, and
reaches that data on demand through `prs.refresh`, which is in the REQUIRED
remote-command set, so no paired build loses PR detail. Lists and badges are
unaffected: slim `pull_requests` rows still sync, and though four iOS
projections name the table as an invalidation trigger, no projection query reads
a column from it. Already-paired devices keep the rows they have, so
previously-opened PRs still render offline. It also ends a scroll-driven write
path — the Lanes page's visible-lane refresh upserts here, so scrolling was
pushing changesets to every phone.

Note for the record: the research report said the `github_pr_*` siblings were
already excluded and this table was merely overlooked. They are not excluded;
the set had no PR tables at all.
